+++ client/src/component/TableHeader.js
@@ -1,5 +1,5 @@
 const getTableHeaders = JSON.parse(localStorage.getItem("JamKelas"));
-export const tableHeader = getTableHeaders.map((item, index) => ({
+export const tableHeader = getTableHeaders?.map((item, index) => ({
   jam: `${item.jam_mulai} - ${item.jam_selesai}`,
   senin: "",
   selasa: "",
