+++ client/src/page/DashboardUser/services/getMyPlan.jsx
@@ -1,12 +1,13 @@
 import { url } from "../../../api/url";
 
-const getMyPlan = async (token) => {
+const getMyPlan = async (token, setIsFilled) => {
   try {
     const response = await url
       .get("/my-plan", {
         headers: { Authorization: `Bearer ${token}` },
       })
       .then((result) => {
+        setIsFilled(true);
         return result;
       })
       .catch((result) => {
