+++ client/src/page/RandomKrs/components/Filter/FilterData.js
@@ -17,7 +17,7 @@ export const List = [
 ];
 
 const getTableHeaders = JSON.parse(localStorage.getItem("JamKelas"));
-export const DataJadwal = getTableHeaders.map((item, index) => ({
+export const DataJadwal = getTableHeaders?.map((item, index) => ({
   Jam: `${item.jam_mulai}-${item.jam_selesai}`,
   Senin: "Senin",
   Selasa: "Selasa",
